removed C# module support to focus on just error msg improvement
