+++ lib/ansible/executor/powershell/module_script_wrapper.ps1
@@ -1,3 +1,6 @@
+# (c) 2018 Ansible Project
+# GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
+
 param(
     [Parameter(Mandatory=$true)][System.Collections.IDictionary]$Payload
 )
