--- test/integration/targets/win_csharp/aliases
@@ -1,2 +1,0 @@
-shippable/windows/group4
-shippable/windows/smoketest
