--- test/integration/targets/win_csharp/tasks/main.yml
@@ -1,100 +1,0 @@
-# TODO: this should be spread out into the async, become and exec targets, for
-# now we will self contain it until the interface is more stable
----
-- name: get current user information
-  win_whoami:
-  register: whoami
-
-- name: test C# module with good exit
-  test_csharp:
-    action: normal
-  register: good
-
-- name: assert test C# module with good exit
-  assert:
-    that:
-    - not good is failed
-    - good is changed
-    - good.sid == whoami.account.sid
-
-- name: test C# module with fail exit
-  test_csharp:
-    action: fail
-  register: fail_action
-  ignore_errors: yes
-
-- name: assert test C# module with fail exit
-  assert:
-    that:
-    - fail_action is failed
-    - fail_action.msg == "failure"
-    - not fail_action.exception is defined
-
-- name: test C# module with uncaught exception
-  test_csharp:
-    action: exception
-  register: exception
-  ignore_errors: yes
-
-- name: assert test C# module with uncaught exception
-  assert:
-    that:
-    - exception is failed
-    - 'exception.msg == "Unhandled exception while executing C# module: exception thrown"'
-    - '"at Ansible.Module.WinCSharp.Main(String[] args)" in exception.exception'
-
-- name: test C# module with uncaught exception in function
-  test_csharp:
-    action: exception_in_function
-  register: exception_in_function
-  ignore_errors: yes
-
-- name: assert test C# module with uncaught exception in function
-  assert:
-    that:
-    - exception_in_function is failed
-    - 'exception_in_function.msg == "Unhandled exception while executing C# module: exception in function"'
-    - '"at Ansible.Module.WinCSharp" in exception_in_function.exception'  # 2008 C@ compiler screws up this line ending so we can't test on the function
-
-- name: test C# module with module_util reference
-  test_csharp_util:
-  register: csharp_util
-
-- name: assert test C# module with module_util reference
-  assert:
-    that:
-    - not csharp_util is failed
-    - not csharp_util is changed
-    - csharp_util.sid == whoami.account.sid
-
-- name: run C# module with become
-  test_csharp:
-    action: normal
-  vars:
-    ansible_become: yes
-    ansible_become_user: SYSTEM
-    ansible_become_method: runas
-  register: csharp_become
-
-- name: assert run C# module with become
-  assert:
-    that:
-    - not csharp_become is failed
-    - csharp_become is changed
-    - csharp_become.action == "normal"
-    - csharp_become.sid == "S-1-5-18"
-    - csharp_become.username == "NT AUTHORITY\\SYSTEM"
-
-- name: run C# module with async
-  test_csharp:
-    action: normal
-  async: 10
-  register: csharp_async
-
-- name: assert run C# module with async
-  assert:
-    that:
-    - not csharp_async is failed
-    - csharp_async is changed
-    - csharp_async.action == "normal"
-    - csharp_async is finished
