+++ test/sanity/code-smell/shebang.py
@@ -22,7 +22,6 @@ def main():
         '': b'#!/usr/bin/python',
         '.py': b'#!/usr/bin/python',
         '.ps1': b'#!powershell',
-        '.cs': b'#!csharp',
     }
 
     skip = set([
